Refactor calendar page into modular components
